chore: move redis config to runtime local
